+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - diagonal and unfold are CPU views
+- public broadcast_to and expand_as are CPU views
 
 ---